Thumbnail styling started
First I thought that the thumbnails should hover when you ride with the cursor over them. It was easy to set with CSS. However, the task is to highlight the thumbnail when displayed in the gallery using js. Now the thumbnail highlights but don't go back after selecting a different image. Need to use the on/off thing or something like that.